fixed possible IndexError by adding condition

+++ backend/database_tools/tools.py
@@ -4,14 +4,17 @@ from post.models import Listing
 
 
 # get most recent listing in database, return in form of python dict
+# return 0 if database is empty
 def get_database_data():
-    # * may want to change getting all data from database to getting only most recent entry
-    posts = Listing.objects.all()  # get all data from database
-    df = pd.DataFrame(posts.values())  # convert to dataframe
-    data = df.to_json(orient='records')  # convert to json
-    data = json.loads(data)  # convert to python dict
-    # return most recent entry in database
-    return data[-1]
+    # check if database contains any data
+    if Listing.objects.exists():
+        posts = Listing.objects.all()  # get all data from database
+        df = pd.DataFrame(posts.values())  # convert to dataframe
+        data = df.to_json(orient='records')  # convert to json
+        data = json.loads(data)  # convert to python dict
+        # return most recent entry in database
+        return data[-1]
+    return 0
 
 
 # remove key from dictionary
